change custom "target" attr to "data-target" to conform with W3C rules
As per w3c specs, in HTML5, customer attrs for tags should be
prefixed with data-. (e.g. <button data-target="a">)

Previosuly, I had a custom "target" attr that was used to make modals
work. Changing this to "data-target" to conform with w3c so that
pages using this will pass w3 validation.